Advanced Software Engineering
 This is all about building Midas system, a high-profile, high-stakes project which is sure to attract the right kind of attention if you can pull it off.
 Midas Core - the component which receives, validates, and records financial transactions. This component depends on quite a few external resources
 - Kafka to receive new transactions, a SQL database to record and validate them, and a REST API to incentivize them. .
Midas Core uses Spring Boot, which should make the job of integrating external resources far easier than it would be with vanilla Java.
Spring Boot - an extension of Spring - provides your application with a robust Dependency Injection framework, and takes care of much of the boilerplate associated with connecting your application to external dependencies. Spring Boot adopts a somewhat opinionated view of how your application should communicate with the aforementioned resources, but in this case, that’s a good thing - it saves us quite a bit of time wiring everything together.